requirements.txt
streamlit==1.39.0
mysql-connector-python==9.0.0
pandas==2.2.3
plotly==5.24.1
python-dateutil==2.9.0.post0
